Add publish option —deps_from_config
This is useful in case of umbrella projects. Package dependencies are
gathered based on rebar.config of each app instead of including all
packages from the umbrella lock file.
This way each package will only depend on the deps of one app.

Also one project app can declare another as dependency. (Project apps
are not included in the lock file so it was not possible to declare
such package dependencies before)